requirement.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
comtypes==1.1.10
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
Pillow==8.2.0
PyAudio==0.2.11
pyngrok==5.0.5
pypiwin32==223
python-pptx==0.6.19
pyttsx3==2.90
pywin32==300
PyYAML==5.4.1
requests==2.25.1
SpeechRecognition==3.8.1
urllib3==1.26.5
Werkzeug==2.0.1
XlsxWriter==1.4.3
